etl/pyproject.toml
[tool.poetry]
name = "etl"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
kafka-python = "^2.0.2"
backoff = "^2.2.1"
python-dotenv = "^1.0.1"
clickhouse-driver = "^0.2.7"
pydantic = "^2.6.4"
pydantic-settings = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
